% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_YKnock.R
\name{create_YKnock}
\alias{create_YKnock}
\title{Second-order Gaussian knockoffs}
\usage{
create_YKnock(Y, method = c("asdp", "equi", "sdp"), shrink = F)
}
\arguments{
\item{Y}{n-by-p matrix of original responses.}

\item{method}{either "equi", "sdp" or "asdp" (default: "asdp").
This determines the method that will be used to minimize the correlation between the original variables and the knockoffs.}

\item{shrink}{whether to shrink the estimated covariance matrix (default: F).}
}
\value{
A n-by-r matrix of knockoff variables.
}
\description{
This function samples second-order multivariate Gaussian knockoff variables.
First, a multivariate Gaussian distribution is fitted to the observations of Y.
Then, Gaussian knockoffs are generated according to the estimated model.
}
\details{
If the argument \code{shrink} is set to T, a James-Stein-type shrinkage estimator for
the covariance matrix is used instead of the traditional maximum-likelihood estimate. This option
requires the package \code{corpcor}. See \code{\link[corpcor]{cov.shrink}} for more details.

Even if the argument \code{shrink} is set to F, in the case that the estimated covariance
matrix is not positive-definite, this function will apply some shrinkage.
}
\references{
Candes et al., Panning for Gold: Model-free Knockoffs for High-dimensional Controlled Variable Selection,
  arXiv:1610.02351 (2016).
  \href{https://web.stanford.edu/group/candes/knockoffs/index.html}{https://web.stanford.edu/group/candes/knockoffs/index.html}
}
\concept{create}
\keyword{internal}
