% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_modelY_classification_coef.R
\name{stat_modelY_classification_coef}
\alias{stat_modelY_classification_coef}
\title{Importance statistics based on a GLM}
\usage{
stat_modelY_classification_coef(X, Y, Yk, nlambda = 100, standardize = TRUE)
}
\arguments{
\item{X}{n-by-p matrix of qualitative predictors.}

\item{Y}{n-by-r matrix of original responses.}

\item{Yk}{n-by-r matrix of knockoff responses.}

\item{nlambda}{The number of lambda values - default is 100.}

\item{standardize}{Logical flag for Y variable standardization. Default is \code{standardize=TRUE}.}
}
\value{
A r-by-2 matrix of statistics \eqn{Z}.
}
\description{
Computes the signed maximum statistic
  \deqn{W_j = \max(Z_j, \tilde{Z}_j) \cdot \mathrm{sgn}(Z_j - \tilde{Z}_j),}
where \eqn{Z_j} and \eqn{\tilde{Z}_j} are the maximum values of
\eqn{\lambda} at which the jth variable and its knockoff, respectively,
enter the generalized linear model.
}
\details{
This function uses the \code{glmnet} package to fit a generalized linear model
via penalized maximum likelihood.

The statistics \eqn{W_j} are constructed by taking the difference
between the coefficient of the j-th responses and its knockoff.
'

The optional \code{nlambda} parameter can be used to control the granularity of the
grid of \eqn{\lambda}'s. The default value of \code{nlambda} is \code{100}.
}
