% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_modelY_maxlam.R
\name{stat_modelY_maxlam}
\alias{stat_modelY_maxlam}
\title{GLM statistics for knockoff}
\usage{
stat_modelY_maxlam(
  X,
  Y,
  Yk,
  generate_lambda = TRUE,
  nlambda = 100,
  standardize = TRUE
)
}
\arguments{
\item{X}{n-by-p matrix of quantitative predictors.}

\item{Y}{n-by-r matrix of original responses.}

\item{Yk}{n-by-r matrix of knockoff responses.}

\item{generate_lambda}{whether to use the method in the YKnock paper to generate lambda sequence. If not, it will use the method of glmnet package to generate lambda.}

\item{nlambda}{The number of lambda values - default is 100.}

\item{standardize}{Logical flag for Y variable standardization. Default is \code{standardize=TRUE}.}
}
\value{
A r-by-2 matrix of statistics \eqn{Z}.
}
\description{
Computes the signed maximum statistic
  \deqn{W_j = \max(Z_j, \tilde{Z}_j) \cdot \mathrm{sgn}(Z_j - \tilde{Z}_j),}
where \eqn{Z_j} and \eqn{\tilde{Z}_j} are the maximum values of
\eqn{\lambda} at which the jth variable and its knockoff, respectively,
enter the generalized linear model.
}
\details{
This function uses \code{glmnet} to compute the regularization path
on a fine grid of \eqn{\lambda}'s.
}
