% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YKnock_filter.R
\name{YKnock_filter}
\alias{YKnock_filter}
\title{The YKncok Filter}
\usage{
YKnock_filter(
  X,
  Y,
  knockoffs = create_YKnock,
  statistic = stat_modelY_coef,
  fdr = 0.1,
  offset = 1
)
}
\arguments{
\item{X}{n-by-p matrix or data frame of predictors.}

\item{Y}{n-by-r matrix or data frame of responses.}

\item{knockoffs}{method used to construct knockoffs for the \eqn{X} variables.
It must be a function taking a n-by-p matrix as input and returning a n-by-p matrix of knockoff variables.
By default, approximate model-X Gaussian knockoffs are used.}

\item{statistic}{statistics used to assess variable importance. By default,
a lasso statistic with cross-validation is used. See the Details section for more information.}

\item{fdr}{target false discovery rate (default: 0.1).}

\item{offset}{either 0 or 1 (default: 1). This is the offset used to compute the rejection threshold on the
statistics. The value 1 yields a slightly more conservative procedure ("knockoffs+") that
controls the false discovery rate (FDR) according to the usual definition,
while an offset of 0 controls a modified FDR.}
}
\value{
An object of class "knockoff.result". This object is a list
 containing at least the following components:
 \item{Y}{matrix of original responses}
 \item{Yk}{matrix of knockoff responses}
 \item{statistic}{computed test statistics}
 \item{threshold}{computed selection threshold}
 \item{selected}{named vector of selected variables}
}
\description{
This function runs the Knockoffs procedure from start to finish, selecting variables
relevant for predicting the outcome of interest.
}
\details{
This function creates the knockoffs, computes the importance statistics,
and selects variables.
It is the main entry point for the knockoff package.



The parameter \code{knockoffs} controls how knockoff variables are created.
By default, the model-Y scenario is assumed and a multivariate normal distribution
is fitted to the original variables \eqn{Y}.


The default importance statistic is \link{stat_modelY_coef}.

It is possible to provide custom functions for the knockoff constructions
or the importance statistics. Some examples can be found in the vignette.
}
\examples{
r = 100;p = 20;n = 40
m = 10  # num of important response
rho=0.2
betaValue = 1.5
SigmaX=matrix(rho,p,p)
diag(SigmaX)=1
betaA=matrix(sample(c(-1,1)*betaValue,size=m*p, replace = TRUE), nrow = m,ncol = p)
beta=matrix(0,r,p)
beta[1:m,]=betaA
sigma=1
X = matrix(rnorm(n*p),n)\%*\%chol(SigmaX)
Y = X \%*\% t(beta) + sqrt(sigma)*matrix(rnorm(n*r),n,r)
result=YKnock_filter(X, Y)
print(result$selected)

# Basic usage with default arguments
result = YKnock_filter (X, Y)
print(result$selected)


}
\references{
Identification of Significant Gene Expression Changes in Multiple Perturbation Experiments using Knockoffs
Tingting Zhao, Guangyu Zhu, Patrick Flaherty
bioRxiv 2021.10.18.464822;
  \href{https://www.biorxiv.org/content/10.1101/2021.10.18.464822v1}{https://www.biorxiv.org/content/10.1101/2021.10.18.464822v1}
}
